requirements.txt
# Excel Commander Backend Dependencies
# Production-ready requirements

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Data Validation
pydantic==2.6.0
pydantic-settings==2.2.0

# AI Integration
openai==1.14.0

# PowerPoint Generation
python-pptx==0.6.23

# Excel/Data Handling
pandas==2.2.0
openpyxl==3.1.2

# Environment & Config
python-dotenv==1.0.1

# Utilities
aiofiles==23.2.1
httpx==0.27.0
